+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+- Support for custom serialization: `SceneReference` now implements the `ISerializable` interface and the corresponding deserialization constructor.
 
 ### Changed